+++ src/app/beer_garden/config.py
@@ -1502,6 +1502,17 @@ _REQUEST_VALIDATION_SPEC = {
     },
 }
 
+_REPLICATION_SPEC = {
+    "type": "dict",
+    "items": {
+        "enabled": {
+            "type": "bool",
+            "default": True,
+            "description": "Publish subset of events to Rabbit to allow all Gardens to stay in sync when replicated",
+        },
+    },
+}
+
 _SPECIFICATION = {
     "auth": _AUTH_SPEC,
     "configuration": _META_SPEC,
@@ -1515,5 +1526,6 @@ _SPECIFICATION = {
     "plugin": _PLUGIN_SPEC,
     "request_validation": _REQUEST_VALIDATION_SPEC,
     "scheduler": _SCHEDULER_SPEC,
+    "replication": _REPLICATION_SPEC,
     "ui": _UI_SPEC,
 }